+++ python/paddle/fluid/tests/unittests/test_auto_parallel_data_unshard.py
@@ -15,164 +15,14 @@
 from __future__ import print_function
 
 import unittest
+import paddle.fluid as fluid
 
-import copy
-import numpy as np
-import random
-
-import paddle
-import paddle.nn as nn
-import paddle.fluid.core as core
-import paddle.distributed.auto_parallel as auto
-import paddle.nn.functional as F
-from paddle.distributed import fleet
-
-paddle.enable_static()
-paddle.distributed.init_parallel_env()
+from test_parallel_dygraph_dataparallel import TestMultipleGpus
 
 
-class TestDataUnReshard(unittest.TestCase):
-    def test_dp2pp1mp1(self):
-        def create_model(train_program, start_program):
-            with paddle.static.program_guard(train_program, start_program):
-
-                ROOT_MESH = auto.ProcessMesh([0, 1])
-                MESH_0 = auto.ProcessMesh([0, 1], ROOT_MESH)
-                input = paddle.static.data(name='input', shape=[2, 8])
-                label = paddle.static.data(name='label', shape=[2, 8])
-
-                weight_attr = paddle.ParamAttr(
-                    initializer=nn.initializer.Normal(
-                        mean=0.0, std=0.02))
-                linear0 = nn.Linear(8, 8, weight_attr)
-                linear1 = nn.Linear(8, 8, weight_attr)
-
-                auto.shard_tensor(input, MESH_0, dim_mapping=[0, -1])
-                auto.shard_tensor(label, MESH_0, dim_mapping=[0, -1])
-                auto.shard_tensor(linear0.weight, MESH_0, dim_mapping=[-1, -1])
-                auto.shard_tensor(linear1.weight, MESH_0, dim_mapping=[-1, -1])
-
-                linear0_out = linear0(input)
-                gelu_out = F.gelu(linear0_out)
-                linear1_out = linear1(gelu_out)
-                error_cost = paddle.nn.functional.square_error_cost(linear1_out,
-                                                                    label)
-                loss = paddle.mean(error_cost)
-                return train_program, start_program, loss, input, label
-
-        train_program = paddle.static.Program()
-        start_program = paddle.static.Program()
-        # serial program
-        train_program, start_program, loss, input, label = create_model(
-            train_program, start_program)
-
-        dist_strategy = fleet.DistributedStrategy()
-        dist_strategy.semi_auto = True
-        fleet.init(is_collective=True, strategy=dist_strategy)
-        optimizer = paddle.fluid.optimizer.AdamOptimizer(
-            learning_rate=0.00001,
-            beta1=0.9,
-            beta2=0.999,
-            epsilon=1e-08,
-            grad_clip=None)
-
-        optimizer = fleet.distributed_optimizer(optimizer)
-        _, _, distributed_startup_program, distributed_main_program = optimizer.minimize(
-            loss, start_program)
-
-        worker_index = paddle.distributed.get_rank()
-        paddle.seed(worker_index + 2021)
-        random.seed(worker_index + 2021)
-        np.random.seed(worker_index + 2021)
-
-        place = paddle.set_device("gpu")
-        exe = paddle.static.Executor(place)
-        exe.run(distributed_startup_program)
-
-        input_data = np.array(range(2 * 8)).reshape([2, 8]).astype("float32")
-        label_data = np.random.randint(0, 10, [2, 8]).astype("float32")
-
-        fetchs = [loss.name, 'input@RESHARD_0']
-        loss_np, shard_data_np = exe.run(
-            distributed_main_program,
-            feed={"input": input_data,
-                  "label": label_data},
-            fetch_list=fetchs)
-        desired = input_data[worker_index].reshape(shard_data_np.shape)
-        np.testing.assert_allclose(shard_data_np, desired)
-
-    def dp1pp1mp2(self):
-        def create_model(train_program, start_program):
-            with paddle.static.program_guard(train_program, start_program):
-
-                ROOT_MESH = auto.ProcessMesh([0, 1])
-                MESH_0 = auto.ProcessMesh([0, 1], ROOT_MESH)
-                input = paddle.static.data(name='input', shape=[8, 8])
-                label = paddle.static.data(name='label', shape=[8, 8])
-
-                weight_attr = paddle.ParamAttr(
-                    initializer=nn.initializer.Normal(
-                        mean=0.0, std=0.02))
-                linear0 = nn.Linear(8, 8, weight_attr)
-                linear1 = nn.Linear(8, 8, weight_attr)
-
-                auto.shard_tensor(input, MESH_0, dim_mapping=[-1, -1])
-                auto.shard_tensor(label, MESH_0, dim_mapping=[-1, -1])
-
-                auto.shard_tensor(linear0.weight, MESH_0, dim_mapping=[-1, 0])
-                auto.shard_tensor(linear1.weight, MESH_0, dim_mapping=[0, -1])
-
-                linear0_out = linear0(input)
-                gelu_out = F.gelu(linear0_out)
-
-                linear1_out = linear1(gelu_out)
-
-                error_cost = paddle.nn.functional.square_error_cost(linear1_out,
-                                                                    label)
-                loss = paddle.mean(error_cost)
-                return train_program, start_program, loss, input, label
-
-        train_program = paddle.static.Program()
-        start_program = paddle.static.Program()
-        # serial program
-        train_program, start_program, loss, input, label = create_model(
-            train_program, start_program)
-
-        dist_strategy = fleet.DistributedStrategy()
-        dist_strategy.semi_auto = True
-        fleet.init(is_collective=True, strategy=dist_strategy)
-        optimizer = paddle.fluid.optimizer.AdamOptimizer(
-            learning_rate=0.00001,
-            beta1=0.9,
-            beta2=0.999,
-            epsilon=1e-08,
-            grad_clip=None)
-
-        optimizer = fleet.distributed_optimizer(optimizer)
-        _, _, distributed_startup_program, distributed_main_program = optimizer.minimize(
-            loss, start_program)
-
-        worker_index = paddle.distributed.get_rank()
-        paddle.seed(worker_index + 2021)
-        random.seed(worker_index + 2021)
-        np.random.seed(worker_index + 2021)
-
-        place = paddle.set_device("gpu")
-        exe = paddle.static.Executor(place)
-        exe.run(distributed_startup_program)
-
-        input_data = np.array(range(8 * 8)).reshape([8, 8]).astype("float32")
-        label_data = np.random.randint(0, 10, [8, 8]).astype("float32")
-
-        fetchs = [loss.name, 'input']
-        loss_np, shard_data_np = exe.run(
-            distributed_main_program,
-            feed={"input": input_data,
-                  "label": label_data},
-            fetch_list=fetchs)
-
-        desired = input_data.reshape(shard_data_np.shape)
-        np.testing.assert_allclose(shard_data_np, desired)
+class TestParallelClassCenterSample(TestMultipleGpus):
+    def test_parallel_class_center_sample(self):
+        self.run_mnist_2gpu('auto_parallel_data_unshard.py')
 
 
 if __name__ == "__main__":
